Closes #89, Closes #90
Description (so this bug never comes up again):
The bytes over the net and received the server (and conversely the client as well) are not sent in discrete packets as evidenced by the protocol, but the length and data contained by the buffer `buf` vary, perhaps in alignment with the actual packet that was set, perhaps not. The data that is received each time `decode` is called is called a "frame." Things work as expected whenever the "frame" is aligned with the length of a packet, but what if the frame contains only the top half of the packet? In fact, what if the frame contains a full packet and another tacked onto the end?

The decoder needed to be rewritten to accommodate for protocol encryption. The decoder thus is able to handle misaligned frames by waiting for the `buf` buffer to be filled by the next invocation of `decode` until there is enough bytes to read a full VarInt or enough bytes to read the full packet as specified in the header, in which then all of the data for a single packet has arrived and may be read safely. The decoder handles extra bytes by setting `buf.readerIndex` to match that which was read, and forces the netty decoder to recall with the same buffer. The decoder then sets the index to correct offset and is able to realign the bytes. The buffer used by decoder to store misaligned frames is released only when a whole packet arrives at the end of the buffer.

However, on top of that, one must also deal with anomalies pertaining to encryption/decryption. Although I am not necessarily an expert on this topic, it turns out that the same set of encrypted data may not be passed to `update(...)` more than once, because the next call to `update(...)` causes the first 16 bytes to be replaced with a new IV (I think?). Either way, the first 16 bytes are lost whenever the decoder must decrypt the same data twice, i.e. when the same set of data is updated with more bytes from the network. My guess is that the IV acts as a seeding value and must correspond with reading the same data as the encrypted Cipher, and thus decrypting the same data must use a new IV since it acts as if it were being incremented with the encryption IV.

Therefore, the buffer must be filled with decrypted bytes, and must always be updated with bytes that are not already in the buffer (as again, the first 16 bytes of same buffer will be replaced with an IV or whatever). The decoder handles this final issue by moving `buf.readerIndex()` to match that of the end of the buffer before reading from `buf` into the buffer. The buffer's `readerIndex` remains the same in order to continue decoding from the old frame. Thus, each frame is only decrypted once and prevents frame corruption via AES replacing the first few bytes with a new IV.