Fixes bossbars, closes TridentSDK/Trident#67.
Cannot use a flag placed in BossBar to determine whether to send packet or not; the first player to call updateBossBars will destroy the STATE variable and therefore prevent the rest of the players from detecting that the BossBar is dirty.

TabLists also only work with different UUIDs, a new TabListItem.Add packet will not change the player's tab list, but it will change their nameplate as expected.